+++ onsiteprint-plugin/CHANGELOG.md
@@ -165,6 +165,7 @@ This document includes the change log notes for the OnsitePrint Plugin - [GitHub
 - Removed the `onclick` in the Add Participant Button, `.op-button-add`.
 - Removed the `onclick` in the Download CSV Button, `.op-button-csv`.
 - Added new Dropdown Element.
+- Added new Variables to the Header.
 
     [EB | header.php](blocks/event/block-template-parts/header.php)